Use same InfoSection component for home and custom pages
