--- front/app/components/LandingPages/citizen/InfoSection.ts
@@ -1,15 +1,0 @@
-import QuillEditedContent from 'components/UI/QuillEditedContent';
-import styled from 'styled-components';
-import { colors } from 'utils/styleUtils';
-
-export const StyledQuillEditedContent = styled(QuillEditedContent)`
-  h1,
-  h2 {
-    color: ${(props) => props.theme.colors.tenantText};
-  }
-
-  p,
-  li {
-    color: ${colors.textSecondary};
-  }
-`;
